Finish README
Don't forget to mention that this is your first time doing this, because you'll let others know that you're a complete beginner noob. 

On GitHub, saved changes are called commits. Each commit has an associated commit message, which is a description explaining why a particular change was made. Commit messages to capture the history of your changes, so other contributors can understand what you’ve done and why.